Library/Mobile Documents/com~apple~CloudDocs/AriesView/BACK/financal-calculation/requirements.txt
flask==2.3.3
flasgger==0.9.7.1
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
pytest==7.4.2
python-dotenv==1.0.0
flask-migrate==4.0.5
firebase-admin==6.2.0
psycopg2-binary==2.9.7
numpy-financial==1.0.0
numpy==1.24.4
pandas==2.0.3
nest-asyncio==1.5.8
